script/requirements_db.txt
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
uuid>=1.30 